requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.5
djangorestframework==3.10.3
idna==2.8
pytz==2020.1
requests==2.22.0
sqlparse==0.3.1
urllib3==1.25.10
